Add TabDevelop component only for development
