+++ src/system.js
@@ -34,6 +34,7 @@ import Selector from "@/components/Selector";
 import Spinner from "@/components/Spinner";
 import SSProSimulator from "@/components/SSProSimulator";
 import Tabs from "@/components/Tabs";
+import TabDevelop from "@/components/TabDevelop";
 import TextInput from "@/components/TextInput";
 import TooltipNotice from "@/components/TooltipNotice";
 import Toggle from "@/components/Toggle";
@@ -67,6 +68,7 @@ const components = [
   Spinner,
   SSProSimulator,
   Tabs,
+  TabDevelop,
   TextInput,
   TooltipNotice,
   Pagination,
@@ -114,6 +116,7 @@ export {
   Spinner,
   SSProSimulator,
   Tabs,
+  TabDevelop,
   TextInput,
   TooltipNotice,
   Pagination,
